V035 19.08.2024: RS : bugfix: buzzer settings not saved, added missing F3F A-Line flyover signal,  long 1.5s finish signals F3F+F3B,                        interpreation of F3F Tasktime is now: time between StartTaskSignal and launch time,                        more precise naming of entities/files: BaseManager, LineController

+++ lib/F3XLib/PinManager.h
@@ -18,7 +18,7 @@ class PinManager {
     void set(bool aState, bool aForce=false) {
       if (myPinState != aState || aForce) {
         myPinState = aState;
-				digitalWrite(myPin, myPinState);
+        digitalWrite(myPin, myPinState);
         // logMsg(INFO, String(F("PinManager::set=")) + String(myPinState)); 
       } 
     }
@@ -75,10 +75,10 @@ class PinManager {
           myStartTime = millis();
           myDuration = aDuration;
           // logMsg(INFO, String(F("PinManager::set():IDLE ")) + String(myStartTime) + "/" + String(myDuration)); 
-				  set(myOnState);
+          set(myOnState);
           break;
         case ON:
-          int16_t remain = myDuration - (millis() - myStartTime);
+          unsigned long remain = myDuration - (millis() - myStartTime);
           if (remain < aDuration ) {
             myStartTime = millis();
             myDuration = aDuration;
@@ -97,9 +97,9 @@ class PinManager {
           if ((now - myStartTime) >= myDuration) {
             // logMsg(INFO, String(F("PinManager::update():ON ->IDLE ")) + String(now) + "/" + String(myStartTime) + "/" + String(myDuration)); 
             myState = IDLE;
-				    set(myOffState);
+            set(myOffState);
           } else {
-				    set(myOnState);
+            set(myOnState);
           }
           break;
         case PATTERN:
